requirements.txt
pygame
neat-python
numpy
networkx
torch
snntorch
hypernetx
configparser
